# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: LabTech GeoLab User Tracking API
  description: |
    Comprehensive API for user activity tracking, authentication, and administrative monitoring in the LabTech GeoLab application.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    - Standard endpoints: 100 requests per minute per IP
    - Authentication endpoints: 10 requests per minute per IP
    
    ## Error Codes
    - 400: Bad Request - Invalid input data
    - 401: Unauthorized - Missing or invalid authentication
    - 403: Forbidden - Insufficient permissions
    - 404: Not Found - Resource doesn't exist
    - 429: Too Many Requests - Rate limit exceeded
    - 500: Internal Server Error - Unexpected server error
    - 503: Service Unavailable - Database connection failed
  version: 1.0.0
  contact:
    name: LabTech GeoLab Support
    email: support@labtech-geolab.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://staging-api.labtech-geolab.com/api/v1
    description: Staging server
  - url: https://api.labtech-geolab.com/api/v1
    description: Production server
...
paths:
...
  /activities:
    get:
      tags:
        - Activities
      summary: Get activity logs
      description: Retrieve paginated activity logs with optional filters
      operationId: getActivities
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: userId
          in: query
          description: Filter by user ID
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          description: Filter by start date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Filter by end date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: action
          in: query
          description: Filter by action type
          schema:
            type: string
            enum: [read, write, delete, open, download]
        - name: resourcePath
          in: query
          description: Filter by resource path (supports wildcards)
          schema:
            type: string
      responses:
        '200':
          description: Activity logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityLogListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Activities
      summary: Log new activity
      description: Create a new activity log entry
      operationId: createActivity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityRequest'
            examples:
              fileRead:
                value:
                  action: read
                  resourceType: file
                  resourcePath: /projects/sample.pdf
                  metadata:
                    fileSize: 1024000
                    duration: 5000
              pageView:
                value:
                  action: open
                  resourceType: page
                  resourcePath: /admin/dashboard
                  metadata:
                    referrer: /home
      responses:
        '201':
          description: Activity logged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityLog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /activities/{id}:
    get:
      tags:
        - Activities
      summary: Get activity by ID
      description: Retrieve a specific activity log entry
      operationId: getActivityById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Activity log ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Activity retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityLog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /activities/export:
    get:
      tags:
        - Activities
      summary: Export activity logs
      description: Export filtered activity logs as CSV or JSON
      operationId: exportActivities
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          required: true
          description: Export format
          schema:
            type: string
            enum: [csv, json]
        - name: startDate
          in: query
          description: Filter by start date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Filter by end date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: userId
          in: query
          description: Filter by user ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Export file generated
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityLog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /activities/stats:
    get:
      tags:
        - Activities
      summary: Get activity statistics
      description: Retrieve aggregated activity statistics
      operationId: getActivityStats
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          description: Start date for statistics (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: End date for statistics (ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/refresh
  schemas:
...
    CreateActivityRequest:
      type: object
      required:
        - action
        - resourceType
        - resourcePath
      properties:
        action:
          type: string
          enum: [read, write, delete, open, download]
        resourceType:
          type: string
          enum: [file, folder, page]
        resourcePath:
          type: string
          example: /projects/sample.pdf
        metadata:
          type: object
          additionalProperties: true
          description: Additional context about the activity
    ActivityLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        username:
          type: string
        action:
          type: string
          enum: [read, write, delete, open, download]
        resourceType:
          type: string
          enum: [file, folder, page]
        resourcePath:
          type: string
        metadata:
          type: object
          additionalProperties: true
        ipAddress:
          type: string
        userAgent:
          type: string
        timestamp:
          type: string
          format: date-time
    ActivityLogListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ActivityLog'
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
    ActivityStats:
      type: object
      properties:
        totalActivities:
          type: integer
        actionBreakdown:
          type: object
          additionalProperties:
            type: integer
        topUsers:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
                format: uuid
              username:
                type: string
              count:
                type: integer
        topFiles:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              count:
                type: integer
        activitiesOverTime:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              count:
                type: integer
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: Invalid request parameters
            details:
              type: object
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              format: uuid
  responses:
    BadRequest:
      description: Bad Request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INVALID_INPUT
              message: Validation failed
              details:
                username: Username must be at least 3 characters
              timestamp: '2024-01-15T10:30:00Z'
              requestId: 550e8400-e29b-41d4-a716-446655440000
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
              timestamp: '2024-01-15T10:30:00Z'
              requestId: 550e8400-e29b-41d4-a716-446655440000
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: Admin access required
              timestamp: '2024-01-15T10:30:00Z'
              requestId: 550e8400-e29b-41d4-a716-446655440000
    NotFound:
      description: Not Found - Resource doesn't exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
              timestamp: '2024-01-15T10:30:00Z'
              requestId: 550e8400-e29b-41d4-a716-446655440000
...